+++ kernel-rs/src/vm.rs
@@ -33,23 +33,23 @@ pub unsafe fn kvminit() {
     ptr::write_bytes(KERNEL_PAGETABLE as *mut libc::CVoid, 0, PGSIZE);
 
     // uart registers
-    kvmmap(UART0, UART0, PGSIZE, (PTE_R | PTE_W) as i32);
+    kvmmap(UART0, UART0, PGSIZE, PTE_R | PTE_W);
 
     // virtio mmio disk interface
-    kvmmap(VIRTIO0, VIRTIO0, PGSIZE, (PTE_R | PTE_W) as i32);
+    kvmmap(VIRTIO0, VIRTIO0, PGSIZE, PTE_R | PTE_W);
 
     // CLINT
-    kvmmap(CLINT, CLINT, 0x10000, (PTE_R | PTE_W) as i32);
+    kvmmap(CLINT, CLINT, 0x10000, PTE_R | PTE_W);
 
     // PLIC
-    kvmmap(PLIC, PLIC, 0x400000, (PTE_R | PTE_W) as i32);
+    kvmmap(PLIC, PLIC, 0x400000, PTE_R | PTE_W);
 
     // map kernel text executable and read-only.
     kvmmap(
         KERNBASE,
         KERNBASE,
         (etext.as_mut_ptr() as usize).wrapping_sub(KERNBASE),
-        (PTE_R | PTE_X) as i32,
+        PTE_R | PTE_X,
     );
 
     // map kernel data and the physical RAM we'll make use of.
@@ -57,7 +57,7 @@ pub unsafe fn kvminit() {
         etext.as_mut_ptr() as usize,
         etext.as_mut_ptr() as usize,
         (PHYSTOP).wrapping_sub(etext.as_mut_ptr() as usize),
-        (PTE_R | PTE_W) as i32,
+        PTE_R | PTE_W,
     );
 
     // map the trampoline for trap entry/exit to
@@ -66,7 +66,7 @@ pub unsafe fn kvminit() {
         TRAMPOLINE,
         trampoline.as_mut_ptr() as usize,
         PGSIZE,
-        (PTE_R | PTE_X) as i32,
+        PTE_R | PTE_X,
     );
 }
 
@@ -95,7 +95,7 @@ unsafe fn walk(mut pagetable: PagetableT, va: usize, alloc: i32) -> *mut PteT {
     }
     for level in (1..3).rev() {
         let pte: *mut PteT = &mut *pagetable.add(px(level, va)) as *mut usize;
-        if *pte & PTE_V as usize != 0 {
+        if *pte & PTE_V != 0 {
             pagetable = pte2pa(*pte) as PagetableT
         } else {
             if alloc == 0 || {
@@ -105,7 +105,7 @@ unsafe fn walk(mut pagetable: PagetableT, va: usize, alloc: i32) -> *mut PteT {
                 return ptr::null_mut();
             }
             ptr::write_bytes(pagetable as *mut libc::CVoid, 0, PGSIZE);
-            *pte = pa2pte(pagetable as usize) | PTE_V as usize
+            *pte = pa2pte(pagetable as usize) | PTE_V
         }
     }
     &mut *pagetable.add(px(0, va)) as *mut usize
@@ -122,7 +122,7 @@ pub unsafe fn walkaddr(pagetable: PagetableT, va: usize) -> usize {
     if pte.is_null() {
         return 0;
     }
-    if *pte & PTE_V as usize == 0 {
+    if *pte & PTE_V == 0 {
         return 0;
     }
     if *pte & PTE_U as usize == 0 {
@@ -151,7 +151,7 @@ pub unsafe fn kvmpa(va: usize) -> usize {
     if pte.is_null() {
         panic(b"kvmpa\x00" as *const u8 as *mut u8);
     }
-    if *pte & PTE_V as usize == 0 {
+    if *pte & PTE_V == 0 {
         panic(b"kvmpa\x00" as *const u8 as *mut u8);
     }
     let pa: usize = pte2pa(*pte);
@@ -176,10 +176,10 @@ pub unsafe fn mappages(
         if pte.is_null() {
             return -1;
         }
-        if *pte & PTE_V as usize != 0 {
+        if *pte & PTE_V != 0 {
             panic(b"remap\x00" as *const u8 as *mut u8);
         }
-        *pte = pa2pte(pa) | perm as usize | PTE_V as usize;
+        *pte = pa2pte(pa) | perm as usize | PTE_V;
         if a == last {
             break;
         }
@@ -201,11 +201,11 @@ pub unsafe fn uvmunmap(pagetable: PagetableT, va: usize, size: usize, do_free: i
         if pte.is_null() {
             panic(b"uvmunmap: walk\x00" as *const u8 as *mut u8);
         }
-        if *pte & PTE_V as usize == 0 {
+        if *pte & PTE_V == 0 {
             printf(b"va=%p pte=%p\n\x00" as *const u8 as *mut u8, a, *pte);
             panic(b"uvmunmap: not mapped\x00" as *const u8 as *mut u8);
         }
-        if pte_flags(*pte) == PTE_V as usize {
+        if pte_flags(*pte) == PTE_V {
             panic(b"uvmunmap: not a leaf\x00" as *const u8 as *mut u8);
         }
         if do_free != 0 {
@@ -245,7 +245,7 @@ pub unsafe fn uvminit(pagetable: PagetableT, src: *mut u8, sz: u32) {
         0,
         PGSIZE,
         mem as usize,
-        (PTE_W | PTE_R | PTE_X | PTE_U) as i32,
+        PTE_W | PTE_R | PTE_X | PTE_U,
     );
     ptr::copy(
         src as *const libc::CVoid,
@@ -274,7 +274,7 @@ pub unsafe fn uvmalloc(pagetable: PagetableT, mut oldsz: usize, newsz: usize) ->
             a,
             PGSIZE,
             mem as usize,
-            (PTE_W | PTE_X | PTE_R | PTE_U) as i32,
+            PTE_W | PTE_X | PTE_R | PTE_U,
         ) != 0
         {
             kfree(mem as *mut libc::CVoid);
@@ -311,8 +311,8 @@ unsafe fn freewalk(pagetable: PagetableT) {
             // this PTE points to a lower-level page table.
             let child: usize = pte2pa(pte);
             freewalk(child as PagetableT);
-            *pagetable.offset(i as isize) = 0
-        } else if pte & PTE_V as usize != 0 {
+            *pagetable.offset(i) = 0
+        } else if pte & PTE_V != 0 {
             panic(b"freewalk: leaf\x00" as *const u8 as *mut u8);
         }
     }
@@ -344,7 +344,7 @@ pub unsafe fn uvmcopy(old: PagetableT, new: PagetableT, sz: usize) -> i32 {
         if pte.is_null() {
             panic(b"uvmcopy: pte should exist\x00" as *const u8 as *mut u8);
         }
-        if *pte & PTE_V as usize == 0 {
+        if *pte & PTE_V == 0 {
             panic(b"uvmcopy: page not present\x00" as *const u8 as *mut u8);
         }
         let pa = pte2pa(*pte);
